hw done

+++ homework/cookies.cpp
@@ -1,21 +1,20 @@
 #include <iostream>
 #include <string>
-#include <iomanip>
 //this program calulates how many ingredients are needed for a recipe
 using namespace std;
 
 int main(){
-    int sugarPerCookie, butterPerCookie, flourPerCookie;
+    double sugarPerCookie, butterPerCookie, flourPerCookie;
     
-    sugarPerCookie = 1.5/48;
-    butterPerCookie = 1/48;  //use float?
-    flourPerCookie = 2.75/48;
+    sugarPerCookie = 1.5/48.0;
+    butterPerCookie = 1/48.0;  //use float?
+    flourPerCookie = 2.75/48.0;
     
     int cookieAmount;
     cout << "How many cookies would you like to make?" << endl;
     cin >> cookieAmount;
 
-    int neededSugar, neededButter, neededFlour;
+    double neededSugar, neededButter, neededFlour;
     neededSugar = cookieAmount * sugarPerCookie;
     neededButter = cookieAmount * butterPerCookie;
     neededFlour = cookieAmount * flourPerCookie;
